Les labels peuvent être placés n'importe où

+++ src/components/memory.nim
@@ -2,10 +2,10 @@ import os, strutils
 
 const BUFFER_SIZE = 4096
 var buffer: array[BUFFER_SIZE, char]
-var size: int = 0
+var size: int64 = 0
 
-proc bufferSize*(): int = size
-proc bufferRead*(index: int): char = buffer[index]
+proc bufferSize*(): int64 = size
+proc bufferRead*(index: int64): char = buffer[index]
 
 proc toMemory*(file: File) =
   while not endOfFile(file) and size < BUFFER_SIZE:
